backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB
pymongo[srv]==4.6.1
motor==3.3.2
beanie==1.25.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[argon2]
argon2-cffi>=21.3.0

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2

# File handling
aiofiles==23.2.1
Pillow==10.2.0

# Excel/CSV
openpyxl==3.1.2
pandas==2.1.4

# Barcode
python-barcode==0.15.1

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
